+++ src/main/java/com/booleanuk/api/cinema/movie/MovieController.java
@@ -1,11 +1,12 @@
 package com.booleanuk.api.cinema.movie;
 
+import com.booleanuk.api.cinema.responses.ErrorResponse;
+import com.booleanuk.api.cinema.responses.Response;
 import com.booleanuk.api.cinema.screening.Screening;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDateTime;
@@ -20,7 +21,17 @@ public class MovieController {
     private MovieRepository repo;
 
     @PostMapping
-    public ResponseEntity<Movie> create(@RequestBody Movie movie){
+    public ResponseEntity<Response<?>> create(@RequestBody Movie movie){
+        if (movie.getTitle() == null ||
+                movie.getRating() == null ||
+                movie.getDescription() == null ||
+                movie.getRuntimeMins() <= 0){
+            ErrorResponse badRequest = new ErrorResponse();
+            badRequest.set("bad request");
+
+            return new ResponseEntity<>(badRequest, HttpStatus.BAD_REQUEST);
+        }
+
         movie.setCreatedAt(nowFormatted());
         movie.setUpdatedAt(nowFormatted());
 
@@ -32,22 +43,55 @@ public class MovieController {
             }
         }
 
-        return new ResponseEntity<Movie>(repo.save(movie), HttpStatus.CREATED);
+        repo.save(movie);
+        Response<Movie> movieResponse = new Response<>();
+        movieResponse.set(movie);
+
+        return new ResponseEntity<>(movieResponse, HttpStatus.CREATED);
     }
 
     @GetMapping
-    public List<Movie> getAll(){
-        return repo.findAll();
+    public ResponseEntity<Response<List<Movie>>> getAll(){
+        Response<List<Movie>> movieListResponse = new Response<>();
+        movieListResponse.set(repo.findAll());
+
+        return ResponseEntity.ok(movieListResponse);
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<Movie> getOne(@PathVariable int id){
-        return ResponseEntity.ok(getById(id));
+    public ResponseEntity<Response<?>> getOne(@PathVariable int id){
+        Movie movie = getById(id);
+
+        if (movie == null){
+            ErrorResponse notFound = new ErrorResponse();
+            notFound.set("not found");
+
+            return new ResponseEntity<>(notFound, HttpStatus.NOT_FOUND);
+        }
+
+        Response<Movie> movieResponse = new Response<>();
+        movieResponse.set(movie);
+
+        return ResponseEntity.ok(movieResponse);
     }
 
     @PutMapping("{id}")
-    public ResponseEntity<Movie> update(@PathVariable int id, @RequestBody Movie movie){
+    public ResponseEntity<Response<?>> update(@PathVariable int id, @RequestBody Movie movie){
+        if (movie.getRuntimeMins() != 0 && movie.getRuntimeMins() <= 0){
+            ErrorResponse badRequest = new ErrorResponse();
+            badRequest.set("bad request");
+
+            return new ResponseEntity<>(badRequest, HttpStatus.BAD_REQUEST);
+        }
+
         Movie toUpdate = getById(id);
+
+        if (toUpdate == null){
+            ErrorResponse notFound = new ErrorResponse();
+            notFound.set("not found");
+
+            return new ResponseEntity<>(notFound, HttpStatus.NOT_FOUND);
+        }
 
         Optional.ofNullable(movie.getTitle())
                 .ifPresent(title -> toUpdate.setTitle(title));
@@ -59,21 +103,34 @@ public class MovieController {
                 .ifPresent(runtimeMins -> toUpdate.setRuntimeMins(runtimeMins));
 
         toUpdate.setUpdatedAt(nowFormatted());
+        repo.save(toUpdate);
+        Response<Movie> movieResponse = new Response<>();
+        movieResponse.set(toUpdate);
 
-        return new ResponseEntity<Movie>(repo.save(toUpdate), HttpStatus.CREATED);
+        return new ResponseEntity<>(movieResponse, HttpStatus.CREATED);
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<Movie> delete(@PathVariable int id){
+    public ResponseEntity<Response<?>> delete(@PathVariable int id){
         Movie toDelete = getById(id);
+
+        if (toDelete == null){
+            ErrorResponse notFound = new ErrorResponse();
+            notFound.set("not found");
+
+            return new ResponseEntity<>(notFound, HttpStatus.NOT_FOUND);
+        }
+
         repo.delete(toDelete);
         toDelete.setScreenings(new ArrayList<Screening>());
+        Response<Movie> movieResponse = new Response<>();
+        movieResponse.set(toDelete);
 
-        return ResponseEntity.ok(toDelete);
+        return ResponseEntity.ok(movieResponse);
     }
 
     private Movie getById(int id){
-        return repo.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        return repo.findById(id).orElse(null);
     }
 
     private String nowFormatted(){
